+++ src/geo/mod.rs
@@ -1,6 +1,7 @@
 pub mod crs;
 pub mod datatypes;
 pub mod functions;
+pub mod geojson;
 pub mod gml;
 pub mod places;
 pub mod spatial_index;